refactor: dissolve markdown god file in index properties
