+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/constants.ts
@@ -57,3 +57,13 @@ export const EMPTY_METADATA: PartitionedFieldMetadata = {
   incompatible: [],
   sameFamily: [],
 };
+
+export const EMPTY_PLACEHOLDER = '--';
+
+export const ECS_FIELD_REFERENCE_URL =
+  'https://www.elastic.co/guide/en/ecs/current/ecs-field-reference.html';
+
+/** The documentation link shown in the `Data Quality` dashboard */
+export const ECS_REFERENCE_URL = 'https://www.elastic.co/guide/en/ecs/current/ecs-reference.html';
+export const MAPPING_URL =
+  'https://www.elastic.co/guide/en/elasticsearch/reference/current/mapping.html';
