+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/index_check_flyout/index_properties/index_check_fields/tabs/compare_fields_table/same_family/index.test.tsx
@@ -8,9 +8,9 @@
 import { render, screen } from '@testing-library/react';
 import React from 'react';
 
-import { SAME_FAMILY } from './translations';
 import { TestExternalProviders } from '../../../../../../../../../mock/test_providers/test_providers';
 import { SameFamily } from '.';
+import { SAME_FAMILY_BADGE_LABEL } from '../../../translate';
 
 describe('SameFamily', () => {
   test('it renders a badge with the expected content', () => {
@@ -20,6 +20,6 @@ describe('SameFamily', () => {
       </TestExternalProviders>
     );
 
-    expect(screen.getByTestId('sameFamily')).toHaveTextContent(SAME_FAMILY);
+    expect(screen.getByTestId('sameFamily')).toHaveTextContent(SAME_FAMILY_BADGE_LABEL);
   });
 });
