+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/index_check_flyout/index_properties/index_check_fields/tabs/custom_tab/helpers.ts
@@ -15,7 +15,7 @@ import {
   getMarkdownTable,
   getTabCountsMarkdownComment,
   getSummaryTableMarkdownComment,
-} from '../../../markdown/helpers';
+} from '../../utils/markdown';
 import * as i18n from '../../../translations';
 import type {
   CustomFieldMetadata,
