+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/index_check_flyout/index_properties/index_check_fields/tabs/same_family_tab/helpers.ts
@@ -17,9 +17,9 @@ import {
   getIncompatibleMappingsMarkdownTableRows,
   getMarkdownComment,
   getMarkdownTable,
-  getSummaryTableMarkdownComment,
   getTabCountsMarkdownComment,
-} from '../../../markdown/helpers';
+  getSummaryTableMarkdownComment,
+} from '../../utils/markdown';
 import * as i18n from '../../../translations';
 import { SAME_FAMILY_FIELD_MAPPINGS_TABLE_TITLE } from './translations';
 import type {
