+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/index_check_flyout/index_properties/index_stats_panel/index.tsx
@@ -9,9 +9,7 @@ import { EuiFlexGroup, EuiFlexItem, EuiPanel, EuiSpacer } from '@elastic/eui';
 import React from 'react';
 import styled from 'styled-components';
 
-import { DOCS } from '../translations';
-import { ILM_PHASE } from '../../../../../../translations';
-import { SIZE } from '../../../summary_table/translations';
+import { DOCS, ILM_PHASE, SIZE } from '../../../../../../translations';
 import { Stat } from '../../../../../../stat';
 import { getIlmPhaseDescription } from '../../../../../../utils/get_ilm_phase_description';
