+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_summary/summary_actions/check_status/errors_popover/errors_viewer/index.test.tsx
@@ -8,9 +8,10 @@
 import { render, screen } from '@testing-library/react';
 import React from 'react';
 
+import { INDEX } from '../../../../../translations';
 import { TestExternalProviders } from '../../../../../mock/test_providers/test_providers';
-import { ERROR, INDEX, PATTERN } from './translations';
 import { ErrorSummary } from '../../../../../types';
+import { ERROR, PATTERN } from '../../../translations';
 import { ErrorsViewer } from '.';
 
 interface ExpectedColumns {
