--- x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_summary/summary_actions/check_status/errors_popover/errors_viewer/translations.ts
@@ -1,29 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { i18n } from '@kbn/i18n';
-
-export const ERROR = i18n.translate(
-  'securitySolutionPackages.ecsDataQualityDashboard.errorsViewerTable.errorColumn',
-  {
-    defaultMessage: 'Error',
-  }
-);
-
-export const INDEX = i18n.translate(
-  'securitySolutionPackages.ecsDataQualityDashboard.errorsViewerTable.indexColumn',
-  {
-    defaultMessage: 'Index',
-  }
-);
-
-export const PATTERN = i18n.translate(
-  'securitySolutionPackages.ecsDataQualityDashboard.errorsViewerTable.patternColumn',
-  {
-    defaultMessage: 'Pattern',
-  }
-);
