+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_summary/summary_actions/index.tsx
@@ -12,16 +12,6 @@ import styled from 'styled-components';
 
 import { CheckAll } from './check_all';
 import { CheckStatus } from './check_status';
-import { ERROR, INDEX, PATTERN } from './check_status/errors_popover/errors_viewer/translations';
-import { ERRORS } from './check_status/errors_popover/translations';
-import {
-  getDataQualitySummaryMarkdownComment,
-  getErrorsMarkdownTable,
-  getErrorsMarkdownTableRows,
-  getPatternSummaryMarkdownComment,
-  getSummaryTableMarkdownHeader,
-  getSummaryTableMarkdownRow,
-} from '../../data_quality_details/indices_details/pattern/index_check_flyout/index_properties/markdown/helpers';
 import type { DataQualityCheckResult, IndexToCheck, PatternRollup } from '../../types';
 import { useDataQualityContext } from '../../data_quality_context';
 import { useResultsRollupContext } from '../../contexts/results_rollup_context';
@@ -30,6 +20,15 @@ import { getErrorSummaries } from './utils/get_error_summaries';
 import { getSizeInBytes } from '../../utils/stats';
 import { getSummaryTableItems } from '../../utils/get_summary_table_items';
 import { defaultSort } from '../../constants';
+import {
+  getDataQualitySummaryMarkdownComment,
+  getErrorsMarkdownTable,
+  getErrorsMarkdownTableRows,
+  getPatternSummaryMarkdownComment,
+} from './utils/markdown';
+import { getSummaryTableMarkdownHeader, getSummaryTableMarkdownRow } from '../../utils/markdown';
+import { ERROR, ERRORS, PATTERN } from './translations';
+import { INDEX } from '../../translations';
 
 const StyledActionsContainerFlexItem = styled(EuiFlexItem)`
   margin-top: auto;
