+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_summary/summary_actions/translations.ts
@@ -1,0 +1,36 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { i18n } from '@kbn/i18n';
+
+export const DATA_QUALITY = i18n.translate(
+  'securitySolutionPackages.ecsDataQualityDashboard.dataQuality',
+  {
+    defaultMessage: 'Data quality',
+  }
+);
+
+export const ERROR = i18n.translate(
+  'securitySolutionPackages.ecsDataQualityDashboard.errors.error',
+  {
+    defaultMessage: 'Error',
+  }
+);
+
+export const ERRORS = i18n.translate(
+  'securitySolutionPackages.ecsDataQualityDashboard.errors.errors',
+  {
+    defaultMessage: 'Errors',
+  }
+);
+
+export const PATTERN = i18n.translate(
+  'securitySolutionPackages.ecsDataQualityDashboard.errors.pattern',
+  {
+    defaultMessage: 'Pattern',
+  }
+);
